♻️ [refactor]#264: 검색 컴포넌트 로직 유틸화
